// comanda-front/src/types/shared/authResponse.ts
// Generated by https://quicktype.io

export interface AuthResponse {
    token: string;
    user:  AuthUser;
}

export interface AuthUser {
    id:        string;
    firstName: string;
    lastName:  string;
    email:     string;
    roles:     string[];
    isActive:  boolean;
    createdAt: string;
    updatedAt: string;
}

export interface PartialUser extends Partial<AuthUser> {}